use size_t for better compatibility

+++ Algorithms.h
@@ -5,106 +5,106 @@
 #include <algorithm>
 
 class Solution {
-	static bool is_linear(const std::vector<std::vector<unsigned long long>>& _coefficients);
+	static bool is_linear(const std::vector<std::vector<size_t>>& _coefficients);
 
 public:
-	const std::vector<std::vector<unsigned long long>>& coefficients;
-	const unsigned long long range;
+	const std::vector<std::vector<size_t>>& coefficients;
+	const size_t range;
 	const bool linear;
 
 	Solution(
-		const std::vector<std::vector<unsigned long long>>& _coefficients,
-		const unsigned long long _range
+		const std::vector<std::vector<size_t>>& _coefficients,
+		const size_t _range
 	);
 
-	std::pair<std::vector<unsigned long long>, std::vector<unsigned long long>> solve();
+	std::pair<std::vector<size_t>, std::vector<size_t>> solve();
 };
 
 // interface for our algorithms
 class Method {
 protected:
-	void reduce(std::vector<unsigned long long>& equation);
+	void reduce(std::vector<size_t>& equation);
 
 	const Solution* data_ptr;
 	void invalid_call() const throw(...);
 public:
 	Method(const Solution* _data);
 	
-	std::vector<unsigned long long> answers;
-	std::vector<unsigned long long> basis;
+	std::vector<size_t> answers;
+	std::vector<size_t> basis;
 
-	virtual std::pair<std::vector<unsigned long long>, std::vector<unsigned long long>> solve();
+	virtual std::pair<std::vector<size_t>, std::vector<size_t>> solve();
 
 };
 
 class Quine : public Method {
 	struct Value {
-		const unsigned long long index;
+		const size_t index;
 		const bool positive;
 
-		Value(unsigned long long _index, bool _positive = true);
+		Value(size_t _index, bool _positive = true);
 
-		unsigned long long count(const unsigned long long coef) const throw (...);
+		size_t count(const size_t coef) const throw (...);
 
-		friend unsigned long long operator+ (const Value& val, const unsigned long long vec);
+		friend size_t operator+ (const Value& val, const size_t vec);
 
 	private:
-		unsigned long long get_number() const;
+		size_t get_number() const;
 	};
-	friend unsigned long long operator+ (const Value& val, const unsigned long long vec);
+	friend size_t operator+ (const Value& val, const size_t vec);
 
 	bool update_coefs(
-		std::vector<std::vector<unsigned long long>>& coefs, 
+		std::vector<std::vector<size_t>>& coefs, 
 		const Value& val
 	);
 
 	void set_value(
-		std::vector<std::vector<unsigned long long>> coefs, 
-		const unsigned long long vec, 
+		std::vector<std::vector<size_t>> coefs, 
+		const size_t vec, 
 		const Value& val
 	);
 
 	// find first unknown variable in current @coefs
-	unsigned long long get_nearest_var(
-		const std::vector<std::vector<unsigned long long>>& coefs
+	size_t get_nearest_var(
+		const std::vector<std::vector<size_t>>& coefs
 	) throw (...);
 
 	// like dfs
 	void step(
-		const std::vector<std::vector<unsigned long long>>& coefs, 
-		unsigned long long vec = 0
+		const std::vector<std::vector<size_t>>& coefs, 
+		size_t vec = 0
 	);
 
 public:
 	// Note: zero coefficient means Positive value for equation
 	Quine(const Solution* solution);
 
-	std::pair<std::vector<unsigned long long>, std::vector<unsigned long long>> solve();
+	std::pair<std::vector<size_t>, std::vector<size_t>> solve();
 };
 
 class TSS : public Method {
-	std::vector<std::vector<unsigned long long>> prepare_coefs(
-		std::vector<std::vector<unsigned long long>> coefs, 
-		const unsigned long long range) throw(...);
+	std::vector<std::vector<size_t>> prepare_coefs(
+		std::vector<std::vector<size_t>> coefs, 
+		const size_t range) throw(...);
 
-	std::vector<unsigned long long> get_start_vectors(const std::vector<unsigned long long>& first, unsigned long long range);
+	std::vector<size_t> get_start_vectors(const std::vector<size_t>& first, size_t range);
 
-	bool multiply(const std::vector<unsigned long long>& equation, const unsigned long long value);
+	bool multiply(const std::vector<size_t>& equation, const size_t value);
 
-	unsigned long long calcul_sum(const std::vector<unsigned long long>& coefs, const unsigned long long vec);
+	size_t calcul_sum(const std::vector<size_t>& coefs, const size_t vec);
 
-	std::vector<unsigned long long> add_equation(
-		const std::vector<unsigned long long>& equation, 
-		const std::vector<unsigned long long>& vectors
+	std::vector<size_t> add_equation(
+		const std::vector<size_t>& equation, 
+		const std::vector<size_t>& vectors
 	);
 
 	// first -> individual, second -> basis
-	std::pair<std::vector<unsigned long long>, std::vector<unsigned long long>> separate_solutions(
-		const std::vector<unsigned long long>& vec
+	std::pair<std::vector<size_t>, std::vector<size_t>> separate_solutions(
+		const std::vector<size_t>& vec
 	);
 
 public:
 	TSS(const Solution* solution);
 
-	std::pair<std::vector<unsigned long long>, std::vector<unsigned long long>> solve();
+	std::pair<std::vector<size_t>, std::vector<size_t>> solve();
 };
