+++ templates.h
@@ -3,6 +3,6 @@
 #include <vector>
 #include <string>
 
-using u_i = unsigned long long;
+using u_i = size_t;
 template<typename T> using V = std::vector<T>;
 using S = std::string;
